+++ packages/rpc-core/src/index.ts
@@ -2,5 +2,4 @@ export * from './lamports';
 export * from './rpc-methods';
 export * from './rpc-subscriptions';
 export * from './stringified-bigint';
-export * from './transaction-signature';
 export * from './unix-timestamp';
