+++ packages/rpc-core/src/rpc-subscriptions/signature-notifications.ts
@@ -1,6 +1,7 @@
+import { TransactionSignature } from '@solana/transactions';
+
 import { Commitment, RpcResponse } from '../rpc-methods/common';
 import { TransactionError } from '../transaction-error';
-import { TransactionSignature } from '../transaction-signature';
 
 type SignatureNotificationsApiNotificationReceived = RpcResponse<Readonly<string>>;
